feat(lifetime): rename component folder as resource
