+++ src/lifetime/mod.rs
@@ -7,11 +7,11 @@ use bevy::prelude::*;
 mod player_spawn;
 mod rope_spawn;
 
-mod component;
+mod resource;
 
 use self::player_spawn::queue_for_player_setup_event;
 use self::rope_spawn::wait_for_player_spawn;
-use self::component::SpawnTrack;
+use self::resource::SpawnTrack;
 
 pub struct LifetimePlugin;
