+++ src/lifetime/resource.rs
